moved existing tests to "annotation" package because it uses annotation implementation approach
